Move INET related functionality to DatagramUnicastInetTest
Make DatagramUnicastTest an abstract class
